Data: Split BaseFormatModelTests into read and read-write test classes
